Apply Luca's fixes for ticket #1229.

+++ numpy/core/src/multiarray/convert_datatype.c
@@ -287,7 +287,7 @@ PyArray_CastTo(PyArrayObject *out, PyArrayObject *mp)
     int simple;
     int same;
     PyArray_VectorUnaryFunc *castfunc = NULL;
-    int mpsize = PyArray_SIZE(mp);
+    intp mpsize = PyArray_SIZE(mp);
     int iswap, oswap;
     NPY_BEGIN_THREADS_DEF;
